reorg files + more vocab testing
