+++ src/scripts/game/score.ts
@@ -1,4 +1,4 @@
-import { get } from "./helpers"
+import { get } from "../helpers"
 
 export default function Score() {
   const scoreEl = get("score");
