+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "outDir": "./built",
     "allowJs": true,
-    "target": "es5"
+    "target": "es5",
+    "resolveJsonModule": true,
   },
   "include": [
     "./src/**/*"
